Separate _sample
